- Moved into Kernel, tested user update, delete and creation.
